Add support for values having ranges
Values found in scientific literature may have ranges, as opposed to
being one value. OptionalRangeType has been added, which allows
QuantityModel to accept either a range in the form of a list (or some
similar object) in addition to just having a number, and QuantityModel
will behave the same way, with addition etc. working as expected.