Add comparison for equality between two QuantityModels
The normal equality operator, ==, tests whether the two objects are
equal, e.g. whether they have the same units, dimensions, etc. However,
there are times when this is not what we are interested in, e.g. when
we have two values which are expressed in different units but are just
different representations for the same value. The is_equal() function
tests for this sort of equality