Change auto parser so compound is no longer a special case
This more general prescription should allow for more flexible parsing
of nested models (and also nicer code).
As a side effect, what were tagged as ‘cem’s are now ‘compound’s in
line with the model’s structure. With this one gone, it is the
intention that all the trees generated by parsers should follow the
form of the models, so that interpret functions can be written
automatically for most models.