Add binding attribute to fields, associated fields, and remove subsets function to ModelList
A number of additions to the model base classes:
-The binding attribute is used to specify that this attribute should be
the same for all submodels contained. For example, all submodels should
have the same compound as the parent model.
-The associated model method allows the user to find an associated
record of a certain type. If multiple records exist, the first one
found during a depth-first search will be returned.
-The _remove_subsets function in ModelList allows for the removal of
any models in the list which are subsets of others, and the strict flag
controls whether only strict subsets are removed or ‘subsets’ which are
duplicates are also removed.