Add is_empty property to models
The is_empty property allows for a user to check easily whether a
record is empty. A record should be considered empty if and only if:

-Each field in it is empty.
-What is empty is decided by each field type, which will return this
using the is_empty(value) method. For example, a 0-length string (“”)
or a 0-length list ([]) are both considered empty, as well as None.

The default behaviour is for any value to be always considered
non-empty by BaseType, and it is expected that this will be overridden
when a user creates a new field type.