Add ignore_when_merging flag and SetType
The ignore_when_merging flag allows for the field to be ignored when
merging. For example, this should be set to True for specifiers as
multiple records can have the same specifier while referring to the
same physical model. In the case that this occurs and
ignore_when_merging is set to True, the records will still be merged
into each other, but both will be kept, so that no extracted
information is wasted.

SetType allows for the native representation of sets in CDE. This is
useful for cases when we would previously have had lists and where it
would have been necessary to check for duplicates