Update with SetType support
The compound model is a great place for SetType to be used, as the
names, roles, and labels should be regarded the same regardless of the
order. This fixes some previous shortcomings in interdependency
resolution which would occur from entities not being merged just
because compound names were in the wrong order.