Add compatibility for the new tokenizing logic.
The new tokenizing logic splits on hyphens and brackets etc. Two tests are added for this situation. These are the same sentences as 2 of the original tests, but now in raw sentence instead of tokens.

New logic of counting tokens and characters are added to abbrev.py. Hyphenated strings are counted as 1 token, hyphens does not contribute to abbreviation length, and in long names hyphens are joined without spaces of either side.